remove yet another osx compiler warning
git-svn-id: file:///home/lennart/svn/public/avahi/trunk@1162 941a03a8-eaeb-0310-b9a0-b1bbd8fe43fe

+++ avahi-daemon/static-services.c
@@ -613,7 +613,7 @@ static int static_service_group_load(StaticServiceGroup *g) {
         }
 
         if (!XML_ParseBuffer(parser, n, n == 0)) {
-            avahi_log_error("XML_ParseBuffer() failed at line %d: %s.\n", XML_GetCurrentLineNumber(parser), XML_ErrorString(XML_GetErrorCode(parser)));
+            avahi_log_error("XML_ParseBuffer() failed at line %d: %s.\n", (int) XML_GetCurrentLineNumber(parser), XML_ErrorString(XML_GetErrorCode(parser)));
             goto finish;
         }
